Implemented CRUD for Product

+++ GeekShop.WebApi/GeekShop.Services/ProductService.cs
@@ -2,9 +2,12 @@
 using GeekShop.Domain;
 namespace GeekShop.Services
 {
-    public interface IProductService : ICRUD
+    public interface IProductService
     {
-
+        void Add(Product product);
+        IEnumerable<Product> GetAll();
+        Product? Get(int id);
+        void Delete(int id);
     }
     public class ProductService : IProductService
     {
@@ -14,18 +17,25 @@ namespace GeekShop.Services
         {
             _repository = repository;
         }
+
         public void Add(Product product)
         {
             _repository.Add(product);
         }
+
         public void Delete(int id)
         {
-            _repository.Delete(id);
+            if(_repository.Get(id) is not null)
+            {
+                _repository.Delete(id);
+            }
         }
-        public Product Get(int id)
+
+        public Product? Get(int id)
         {
             return _repository.Get(id);
         }
+
         public IEnumerable<Product> GetAll()
         {
             return _repository.GetAll();
